distilbert-sms-spam-classifier/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
transformers
google-cloud-storage
torch==2.2.2+cpu
